refactor(calendar): share event color helpers for day/week drag preview
