+++ components/app/views/day-view.tsx
@@ -8,6 +8,10 @@ import { cn } from '@/lib/utils'
 import type { CalendarEvent } from '../calendar'
 import { translations, type Language } from '@/lib/i18n'
 import { formatSelectionRange } from '@/components/app/views/selection-range'
+import {
+  getEventAccentColor,
+  getEventBackgroundColor,
+} from '@/components/app/views/event-colors'
 
 const ContextMenu = ({ children }: { children: React.ReactNode }) => (
   <>{children}</>
@@ -138,38 +142,6 @@ export default function DayView({
       timeZone: timezone,
     }
     return new Intl.DateTimeFormat(language, options).format(date)
-  }
-
-  function getDarkerColorClass(color: string) {
-    const colorMapping: Record<string, string> = {
-      'bg-[#E6F6FD]': '#3B82F6',
-      'bg-[#E7F8F2]': '#10B981',
-      'bg-[#FEF5E6]': '#F59E0B',
-      'bg-[#FFE4E6]': '#EF4444',
-      'bg-[#F3EEFE]': '#8B5CF6',
-      'bg-[#FCE7F3]': '#EC4899',
-      'bg-[#EEF2FF]': '#6366F1',
-      'bg-[#FFF0E5]': '#FB923C',
-      'bg-[#E6FAF7]': '#14B8A6',
-    }
-
-    return colorMapping[color] || '#3A3A3A'
-  }
-
-  function getEventBackgroundColor(color: string) {
-    if (!isDark) return undefined
-
-    const darkModeColorMapping: Record<string, string> = {
-      'bg-[#E6F6FD]': '#2F4655',
-      'bg-[#E7F8F2]': '#2D4935',
-      'bg-[#FEF5E6]': '#4F3F1B',
-      'bg-[#FFE4E6]': '#6C2920',
-      'bg-[#F3EEFE]': '#483A63',
-      'bg-[#FCE7F3]': '#5A334A',
-      'bg-[#E6FAF7]': '#1F4A47',
-    }
-
-    return darkModeColorMapping[color]
   }
 
   const isAllDayEvent = (event: CalendarEvent) => {
@@ -523,7 +495,7 @@ export default function DayView({
               left: '0',
               right: '0',
               opacity: isDark ? 1 : 0.9,
-              backgroundColor: getEventBackgroundColor(event.color),
+              backgroundColor: getEventBackgroundColor(event.color, isDark),
               zIndex: 10 + index,
             }}
             onMouseDown={(e) => handleEventDragStart(event, e)}
@@ -544,11 +516,11 @@ export default function DayView({
           >
             <div
               className={cn('absolute left-0 top-0 w-1 h-full rounded-l-md')}
-              style={{ backgroundColor: getDarkerColorClass(event.color) }}
+              style={{ backgroundColor: getEventAccentColor(event.color) }}
             />
             <div
               className="pl-1.5 truncate"
-              style={{ color: getDarkerColorClass(event.color) }}
+              style={{ color: getEventAccentColor(event.color) }}
             >
               {event.title}
             </div>
@@ -626,18 +598,21 @@ export default function DayView({
           left: '2px',
           zIndex: 100,
           border: '2px dashed white',
-          backgroundColor: getEventBackgroundColor(draggingEvent.color),
+          backgroundColor: getEventBackgroundColor(
+            draggingEvent?.color,
+            isDark,
+          ),
           pointerEvents: 'none',
         }}
       >
         <div
           className={cn('absolute left-0 top-0 w-1 h-full rounded-l-md')}
-          style={{ backgroundColor: getDarkerColorClass(draggingEvent.color) }}
+          style={{ backgroundColor: getEventAccentColor(draggingEvent.color) }}
         />
         <div className="pl-1">
           <div
             className="font-medium truncate"
-            style={{ color: getDarkerColorClass(draggingEvent.color) }}
+            style={{ color: getEventAccentColor(draggingEvent.color) }}
           >
             {draggingEvent.title}
           </div>
@@ -770,7 +745,10 @@ export default function DayView({
                       top: `${startMinutes}px`,
                       height: `${height}px`,
                       opacity: isDark ? 1 : 0.9,
-                      backgroundColor: getEventBackgroundColor(event.color),
+                      backgroundColor: getEventBackgroundColor(
+                        event.color,
+                        isDark,
+                      ),
                       width,
                       left,
                       zIndex: column + 1,
@@ -791,14 +769,14 @@ export default function DayView({
                         'absolute left-0 top-0 w-1 h-full rounded-l-md',
                       )}
                       style={{
-                        backgroundColor: getDarkerColorClass(event.color),
+                        backgroundColor: getEventAccentColor(event.color),
                       }}
                     />
                     <div className="pl-1">
                       <div
                         className="font-medium leading-tight break-words"
                         style={{
-                          color: getDarkerColorClass(event.color),
+                          color: getEventAccentColor(event.color),
                           display: '-webkit-box',
                           WebkitBoxOrient: 'vertical',
                           WebkitLineClamp: Math.max(
@@ -814,7 +792,7 @@ export default function DayView({
                       {height >= 40 && (
                         <div
                           className="text-xs truncate"
-                          style={{ color: getDarkerColorClass(event.color) }}
+                          style={{ color: getEventAccentColor(event.color) }}
                         >
                           {formatDateWithTimezone(start)} -{' '}
                           {formatDateWithTimezone(end)}
